pyproject.toml
[project]
name = "pp02"
version = "0.1.0"
description = "Multifunctional LLM-Powered API Service with FastAPI, Meilisearch, and OpenAI Integration"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.115.12",
    "loguru>=0.7.3",
    "meilisearch>=0.34.1",
    "nest-asyncio>=1.6.0",
    "openai>=1.78.0",
    "pydantic>=2.11.4",
    "pydantic-settings>=2.9.1",
    "pyyaml>=6.0.2",
    "uvicorn>=0.34.2",
]
